returned glob_ext_files
